Add automatic copper balancing to the fab-panel panelizer
The gutters between placed assembly panels are balanced to their aggregate
copper density using the board-array density model and spatial solver; the
reserved process margins stay bare and are excluded from the density domain.
The IPC adapter pieces shared by both panelizers (plan/report types, the
spatial solve, feature lowering, and the generated LayerFeature serializer)
move to crate-level copper_balance and generated modules.
