+++ crates/pcb-ipc2581-tools/README.md
@@ -22,7 +22,9 @@ Run `pcb ipc2581 <command> --help` for arguments and output options.
 
 The mathematical and geometric strategy for automatically balancing board
 arrays is documented in
-[`docs/board-array-copper-balancing.md`](docs/board-array-copper-balancing.md).
+[`docs/board-array-copper-balancing.md`](docs/board-array-copper-balancing.md);
+the fabrication-panel pass that extends it is documented in
+[`docs/fab-panel-copper-balancing.md`](docs/fab-panel-copper-balancing.md).
 
 `edit bom` modifies the input file when `--output` is omitted. Specify an output
 path when the source document must remain unchanged.
