+++ crates/pcb-ir/src/dialects/ipc/mod.rs
@@ -32,7 +32,7 @@ pub use balancing_region::{
     ClearanceCertificate, DEFAULT_BALANCING_CLEARANCE_MM, DEFAULT_BALANCING_GAP_RADIUS_MM,
     DEFAULT_BALANCING_NUMERICAL_GUARD_MM, DEFAULT_BALANCING_REGULARIZATION_RADIUS_MM,
     board_array_balancing_region, collect_board_array_balancing_input,
-    inspect_board_array_balancing_input,
+    collect_fab_panel_balancing_input, inspect_board_array_balancing_input,
 };
 pub use document::{Document, Layer};
 pub use feature::{
